HOF unskip + snapshot cluster: 5,737/7 conformance (only fn:transform remains)
- fn:snapshot matches spec-equivalent impl: identity/parentage preserved through
  typed templates (placeholder accumulator), function results, and graft copies;
  deep-equal excludes xmlns; namespace-node() match pattern; non-node items passthrough
- xsl:element suspends sequence accumulator during content construction
- fn:concat/fn:compare#2 anyAtomicType params pass-through; dynamic-call String
  conversion back to spec (XPTY0004 for substring-before(?, 2))
- fn:min/max return xs:integer for all-integer input; system-property Q{} + HOF flag
- xsl:function cache attr; reserved namespaces incl. map/math/array; TVT simple-content
  rules; section 4.3 whitespace stripping; missing F&O registrations
- W3C XSLT 3.0: 5,737 passed / 7 failed / 8,856 skipped (was 5,732/12)

+++ tests/Bosak.XPath.Core.Tests/EndToEndTests.cs
@@ -16,6 +16,7 @@
 //                      | Charles Korthout | 0.4   | 19-05-2026     | Added tests for Map, Array, and Lookup                                                 |
 //                      | Charles Korthout | 0.5   | 13-06-2026     | Updated FunctionCall_Sum to expect xs:integer for all-integer input                    |
 //                      | Charles Korthout | 0.6   | 25-06-2026     | Use general comparison operators for untyped price nodes (aligns with value-comparison semantics) |
+//                      | Charles Korthout | 0.7   | 14-07-2026     | FunctionCall_Min/Max assert xs:integer (spec-correct after MinMax integer preservation) |
 //                      |==================|=======|================|=========================================================================================
 // ===========================================================================================================================================================
 
@@ -645,15 +646,17 @@ public class EndToEndTests
     [Fact]
     public void FunctionCall_Min()
     {
+        // F&O: fn:min over all-integer input returns xs:integer (least common type).
         var result = Evaluate("min((3, 1, 2))", LoadDocument());
-        Assert.Equal(1m, result.DecimalValue);
+        Assert.Equal(1L, result.IntegerValue);
     }
 
     [Fact]
     public void FunctionCall_Max()
     {
+        // F&O: fn:max over all-integer input returns xs:integer (least common type).
         var result = Evaluate("max((3, 1, 2))", LoadDocument());
-        Assert.Equal(3m, result.DecimalValue);
+        Assert.Equal(3L, result.IntegerValue);
     }
 
     [Fact]
